Refactor q_merge for readability and efficiency
The naming such as "qc" and "tmp" is not clear, so I renamed them to
"first_qc" and "cur_qc" respectively. This makes the code easier to
understand.

The loop structure is also simplified by removing the redundant check
for the first node. This is done by explicitly start the loop from the
second node, replacing the list_for_each_entry macro.

Change-Id: I126a417e84fa8fba8b1581f66632a983ceac0037