Fix duplicate requests and handle empty commands
The original code had an issue where browsers sent duplicate requests
due to the shortcut icon link 'href="#"'. This was fixed by changing
the 'href="#"' to an empty data URL
'href="data:image/x-icon;", type="image/x-icon"' to prevent unintended
requests.

After the fix, the terminal will recieve a '.'. This is because the
browser sends an empty request. To handle this, a new command '.' was
added. This command logs "Empty command" when '.' command is provided.
'.' was registered as a command in 'init_cmd()' to prevent "Unknown
command '.'" errors.

Change-Id: Ie9c5a49693bee1fdc295a1982d747dca4bfa012b